Core: Partial redesign of the audio portion of the core.
There is now a high quality sampling mode that was made possible by the decimator found in the nintaco source code. It runs about half as fast but sounds lovely.

There is now the ability to pan different audio channels between different Left and Right outputs. Not completed but getting there.

The core now builds a buffer of audio samples then sends them to the UI all at once. This utilizes the send audio frame method from the callback.

The AudioMixer class once again is being used and is handling all the heavy lifting.

There is a net decrease in perfomace for this commit. I'll be looking into increasing performance.

There is a bunch of clean up that still needs to happen. But I'm happy enough with the state of things to get it commited.